api/requirements.txt
request
assertpy
pytest